Rename mod season to types and refactor season check.
