+++ pallets/ajuna-awesome-avatars/src/types.rs
